requirements.txt
fastapi
uvicorn[standard]

sqlalchemy[asyncio]
alembic
asyncpg
psycopg2-binary

pydantic
python-dotenv
regex
